SD splits & cleanup
