+++ src/SD/se_tbl.c
@@ -1,236 +1,5 @@
 #include "SD/sound.h"
-
-void no_cmd_80087A80(void);
-void tempo_set_800873CC(void);
-void tempo_move_800873E4(void);
-void sno_set_80086E38(void);
-void svl_set_80086E78(void);
-void svp_set_80086EB8(void);
-void vol_chg_8008756C(void);
-void vol_move_8008758C(void);
-void ads_set_80087904(void);
-void srs_set_8008798C(void);
-void rrs_set_800879E4(void);
-void pan_set_80086F00(void);
-void pan_move_80086F50(void);
-void trans_set_8008750C(void);
-void detune_set_80087730(void);
-void vib_set_80087018(void);
-void vib_change_80087120(void);
-void rdm_set_8008716C(void);
-void swp_set_8008774C(void);
-void sws_set_800876D4(void);
-void por_set_80087670(void);
-void lp1_start_800871B4(void);
-void lp1_end_800871E0(void);
-void lp2_start_800872C0(void);
-void lp2_end_800872EC(void);
-void l3s_set_8008736C(void);
-void l3e_set_80087384(void);
-void kakko_start_80087834(void);
-void kakko_end_80087854(void);
-void use_set_80086EF8(void);
-void rest_set_80086D18(void);
-void tie_set_80086D9C(void);
-void echo_set1_80087754(void);
-void echo_set2_8008775C(void);
-void eon_set_80087764(void);
-void eof_set_800877CC(void);
-void block_end_80087A58(void);
-
-int dword_8009F7B4 = 0xFFFFFFFF;
-
-char *dword_8009F7B8 = 0;
-
-/***** sd_sub1.c *****/
-
-void (*cntl_tbl_8009F7BC[128])(void) = {
-    /* 0x00 */ no_cmd_80087A80,
-    /* 0x01 */ no_cmd_80087A80,
-    /* 0x02 */ no_cmd_80087A80,
-    /* 0x03 */ no_cmd_80087A80,
-    /* 0x04 */ no_cmd_80087A80,
-    /* 0x05 */ no_cmd_80087A80,
-    /* 0x06 */ no_cmd_80087A80,
-    /* 0x07 */ no_cmd_80087A80,
-    /* 0x08 */ no_cmd_80087A80,
-    /* 0x09 */ no_cmd_80087A80,
-    /* 0x0a */ no_cmd_80087A80,
-    /* 0x0b */ no_cmd_80087A80,
-    /* 0x0c */ no_cmd_80087A80,
-    /* 0x0d */ no_cmd_80087A80,
-    /* 0x0e */ no_cmd_80087A80,
-    /* 0x0f */ no_cmd_80087A80,
-    /* 0x10 */ no_cmd_80087A80,
-    /* 0x11 */ no_cmd_80087A80,
-    /* 0x12 */ no_cmd_80087A80,
-    /* 0x13 */ no_cmd_80087A80,
-    /* 0x14 */ no_cmd_80087A80,
-    /* 0x15 */ no_cmd_80087A80,
-    /* 0x16 */ no_cmd_80087A80,
-    /* 0x17 */ no_cmd_80087A80,
-    /* 0x18 */ no_cmd_80087A80,
-    /* 0x19 */ no_cmd_80087A80,
-    /* 0x1a */ no_cmd_80087A80,
-    /* 0x1b */ no_cmd_80087A80,
-    /* 0x1c */ no_cmd_80087A80,
-    /* 0x1d */ no_cmd_80087A80,
-    /* 0x1e */ no_cmd_80087A80,
-    /* 0x1f */ no_cmd_80087A80,
-    /* 0x20 */ no_cmd_80087A80,
-    /* 0x21 */ no_cmd_80087A80,
-    /* 0x22 */ no_cmd_80087A80,
-    /* 0x23 */ no_cmd_80087A80,
-    /* 0x24 */ no_cmd_80087A80,
-    /* 0x25 */ no_cmd_80087A80,
-    /* 0x26 */ no_cmd_80087A80,
-    /* 0x27 */ no_cmd_80087A80,
-    /* 0x28 */ no_cmd_80087A80,
-    /* 0x29 */ no_cmd_80087A80,
-    /* 0x2a */ no_cmd_80087A80,
-    /* 0x2b */ no_cmd_80087A80,
-    /* 0x2c */ no_cmd_80087A80,
-    /* 0x2d */ no_cmd_80087A80,
-    /* 0x2e */ no_cmd_80087A80,
-    /* 0x2f */ no_cmd_80087A80,
-    /* 0x30 */ no_cmd_80087A80,
-    /* 0x31 */ no_cmd_80087A80,
-    /* 0x32 */ no_cmd_80087A80,
-    /* 0x33 */ no_cmd_80087A80,
-    /* 0x34 */ no_cmd_80087A80,
-    /* 0x35 */ no_cmd_80087A80,
-    /* 0x36 */ no_cmd_80087A80,
-    /* 0x37 */ no_cmd_80087A80,
-    /* 0x38 */ no_cmd_80087A80,
-    /* 0x39 */ no_cmd_80087A80,
-    /* 0x3a */ no_cmd_80087A80,
-    /* 0x3b */ no_cmd_80087A80,
-    /* 0x3c */ no_cmd_80087A80,
-    /* 0x3d */ no_cmd_80087A80,
-    /* 0x3e */ no_cmd_80087A80,
-    /* 0x3f */ no_cmd_80087A80,
-    /* 0x40 */ no_cmd_80087A80,
-    /* 0x41 */ no_cmd_80087A80,
-    /* 0x42 */ no_cmd_80087A80,
-    /* 0x43 */ no_cmd_80087A80,
-    /* 0x44 */ no_cmd_80087A80,
-    /* 0x45 */ no_cmd_80087A80,
-    /* 0x46 */ no_cmd_80087A80,
-    /* 0x47 */ no_cmd_80087A80,
-    /* 0x48 */ no_cmd_80087A80,
-    /* 0x49 */ no_cmd_80087A80,
-    /* 0x4a */ no_cmd_80087A80,
-    /* 0x4b */ no_cmd_80087A80,
-    /* 0x4c */ no_cmd_80087A80,
-    /* 0x4d */ no_cmd_80087A80,
-    /* 0x4e */ no_cmd_80087A80,
-    /* 0x4f */ no_cmd_80087A80,
-    /* 0x50 */ tempo_set_800873CC,
-    /* 0x51 */ tempo_move_800873E4,
-    /* 0x52 */ sno_set_80086E38,
-    /* 0x53 */ svl_set_80086E78,
-    /* 0x54 */ svp_set_80086EB8,
-    /* 0x55 */ vol_chg_8008756C,
-    /* 0x56 */ vol_move_8008758C,
-    /* 0x57 */ ads_set_80087904,
-    /* 0x58 */ srs_set_8008798C,
-    /* 0x59 */ rrs_set_800879E4,
-    /* 0x5a */ no_cmd_80087A80,
-    /* 0x5b */ no_cmd_80087A80,
-    /* 0x5c */ no_cmd_80087A80,
-    /* 0x5d */ pan_set_80086F00,
-    /* 0x5e */ pan_move_80086F50,
-    /* 0x5f */ trans_set_8008750C,
-    /* 0x60 */ detune_set_80087730,
-    /* 0x61 */ vib_set_80087018,
-    /* 0x62 */ vib_change_80087120,
-    /* 0x63 */ rdm_set_8008716C,
-    /* 0x64 */ swp_set_8008774C,
-    /* 0x65 */ sws_set_800876D4,
-    /* 0x66 */ por_set_80087670,
-    /* 0x67 */ lp1_start_800871B4,
-    /* 0x68 */ lp1_end_800871E0,
-    /* 0x69 */ lp2_start_800872C0,
-    /* 0x6a */ lp2_end_800872EC,
-    /* 0x6b */ l3s_set_8008736C,
-    /* 0x6c */ l3e_set_80087384,
-    /* 0x6d */ kakko_start_80087834,
-    /* 0x6e */ kakko_end_80087854,
-    /* 0x6f */ no_cmd_80087A80,
-    /* 0x70 */ no_cmd_80087A80,
-    /* 0x71 */ use_set_80086EF8,
-    /* 0x72 */ rest_set_80086D18,
-    /* 0x73 */ tie_set_80086D9C,
-    /* 0x74 */ echo_set1_80087754,
-    /* 0x75 */ echo_set2_8008775C,
-    /* 0x76 */ eon_set_80087764,
-    /* 0x77 */ eof_set_800877CC,
-    /* 0x78 */ no_cmd_80087A80,
-    /* 0x79 */ no_cmd_80087A80,
-    /* 0x7a */ no_cmd_80087A80,
-    /* 0x7b */ no_cmd_80087A80,
-    /* 0x7c */ no_cmd_80087A80,
-    /* 0x7d */ no_cmd_80087A80,
-    /* 0x7e */ no_cmd_80087A80,
-    /* 0x7f */ block_end_80087A58
-};
-
-char rdm_tbl_8009F9BC[129] = {
-    159,  60, 178,  82, 175,  69, 199, 137,
-     16, 127, 224, 157, 220,  31,  97,  22,
-     57, 201, 156, 235,  87,   8, 102, 248,
-     90,  36, 191,  14,  62,  21,  75, 219,
-    171, 245,  49,  12,  67,   2,  85, 222,
-     65, 218, 189, 174,  25, 176,  72,  87,
-    186, 163,  54,  11, 249, 223,  23, 168,
-      4,  12, 224, 145,  24,  93, 221, 211,
-     40, 138, 242,  17,  89, 111,   6,  10,
-     52,  42, 121, 172,  94, 167, 131, 198,
-     57, 193, 180,  58,  63, 254,  79, 239,
-     31,   0,  48, 153,  76,  40, 131, 237,
-    138,  47,  44, 102,  63, 214, 108, 183,
-     73,  34, 188, 101, 250, 207,   2, 177,
-     70, 240, 154, 215, 226,  15,  17, 197,
-    116, 246, 122,  44, 143, 251,  25, 106,
-    229
-};
-
-char VIBX_TBL_8009FA40[32] = {
-      0, 32,   56,  80, 104, 128, 144, 160,
-    176, 192, 208, 224, 232, 240, 240, 248,
-    255, 248, 244, 240, 232, 224, 208, 192,
-    176, 160, 144, 128, 104,  80,  56,  32
-};
-
-/***** sd_ioset.c *****/
-
-int pant_8009FA60[41] = {
-    0,   2,   4,   7,   10,  13,  16,  20,  24,  28,  32,  36,  40,  45,
-    50,  55,  60,  65,  70,  75,  80,  84,  88,  92,  96,  100, 104, 107,
-    110, 112, 114, 116, 118, 120, 122, 123, 124, 125, 126, 127, 127
-};
-
-int se_pant_8009FB04[65] = {
-    0,   2,   4,   6,   8,   10,  14,  18,  22,  28,  34,  40,  46,
-    52,  58,  64,  70,  76,  82,  88,  94,  100, 106, 112, 118, 124,
-    130, 136, 142, 148, 154, 160, 166, 172, 178, 183, 188, 193, 198,
-    203, 208, 213, 217, 221, 224, 227, 230, 233, 236, 238, 240, 242,
-    244, 246, 248, 249, 250, 251, 252, 253, 254, 254, 255, 255, 255
-};
-
-int freq_tbl_8009FC08[108] = {
-    0x10B,  0x11B,  0x12C,  0x13E,  0x151,  0x165,  0x17A,  0x191,  0x1A9,  0x1C2,  0x1DD,  0x1F9,
-    0x217,  0x237,  0x259,  0x27D,  0x2A3,  0x2CB,  0x2F5,  0x322,  0x352,  0x385,  0x3BA,  0x3F3,
-    0x42F,  0x46F,  0x4B2,  0x4FA,  0x546,  0x596,  0x5EB,  0x645,  0x6A5,  0x70A,  0x775,  0x7E6,
-    0x85F,  0x8DE,  0x965,  0x9F4,  0xA8C,  0xB2C,  0xBD6,  0xC8B,  0xD4A,  0xE14,  0xEEA,  0xFCD,
-    0x10BE, 0x11BD, 0x12CB, 0x13E9, 0x1518, 0x1659, 0x17AD, 0x1916, 0x1A94, 0x1C28, 0x1DD5, 0x1F9B,
-    0x217C, 0x237A, 0x2596, 0x27D2, 0x2A30, 0x2CB2, 0x2F5A, 0x322C, 0x3528, 0x3850, 0x3BAC, 0x3F36,
-    0x21,   0x23,   0x26,   0x28,   0x2A,   0x2D,   0x2F,   0x32,   0x35,   0x38,   0x3C,   0x3F,
-    0x42,   0x46,   0x4B,   0x4F,   0x54,   0x59,   0x5E,   0x64,   0x6A,   0x70,   0x77,   0x7E,
-    0x85,   0x8D,   0x96,   0x9F,   0xA8,   0xB2,   0xBD,   0xC8,   0xD4,   0xE1,   0xEE,   0xFC
-};
-
-/***** se_tbl.c *****/
+#include "SD/sd_incl.h"
 
 // TODO: Convert each SE track in the blob back to individual, named arrays.
 //
@@ -247,7 +16,7 @@ int freq_tbl_8009FC08[108] = {
 // Each track has a basename (limited to 8-chars), "jingle03" in this case.
 // The double-digit suffix indicates the index into SETBL.addr[].
 
-int se_data_blob_8009FDB8[ 2371 ] = {
+unsigned int se_data_blob_8009FDB8[ 2370 ] = {
     /* se_data## */
     0xD0780000, 0xD5FF0000, 0xD2290000, 0xD77F0F08,
     0xD8440000, 0xD9100000, 0x2B095A40, 0xFFFE0000,
@@ -1103,446 +872,141 @@ int se_data_blob_8009FDB8[ 2371 ] = {
     0xE00E6356, 0x220A6362, 0xD762000F, 0x240F632C,
     0xFFFE0000,
 #endif
-    /* se_dummy */
-    0xFFFE0000
 };
+
+unsigned char se_dummy[] = { 0x00,0x00,0xfe,0xff };
 
 #define SE_BUF(off)     ((int)&se_data_blob_8009FDB8[off])
-#define SE_DUMMY        SE_BUF(sizeof(se_data_blob_8009FDB8) / sizeof(*se_data_blob_8009FDB8) - 1)
+#define SE_DUMMY        ((int)&se_dummy[0])
 
 SETBL se_tbl_800A22C4[128] = {
-    { 0x01, 0x01, 0x01, 0x00, { SE_DUMMY, SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(561), SE_DUMMY, SE_DUMMY } },
-    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(0), SE_DUMMY, SE_DUMMY } },
-    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(8), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(804), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(2122), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(2132), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1111), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1142), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(1159), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x00, 0x00, { SE_BUF(1171), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1376), SE_DUMMY, SE_DUMMY } },
-    { 0xB0, 0x01, 0x00, 0x00, { SE_BUF(1413), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1398), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1612), SE_DUMMY, SE_DUMMY } },
-    { 0xF0, 0x03, 0x00, 0x00, { SE_BUF(1428), SE_BUF(1441), SE_BUF(1452) } },
-    { 0x20, 0x01, 0x00, 0x00, { SE_BUF(406), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(430), SE_DUMMY, SE_DUMMY } },
-    { 0xA0, 0x01, 0x01, 0x01, { SE_BUF(1921), SE_DUMMY, SE_DUMMY } },
-    { 0xA0, 0x01, 0x01, 0x01, { SE_BUF(571), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x01, 0x00, { SE_BUF(693), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x00, 0x00, { SE_BUF(450), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(464), SE_BUF(476), SE_DUMMY } },
-    { 0x10, 0x01, 0x00, 0x00, { SE_BUF(498), SE_DUMMY, SE_DUMMY } },
-    { 0xA0, 0x01, 0x01, 0x01, { SE_BUF(518), SE_DUMMY, SE_DUMMY } },
-    { 0xA0, 0x01, 0x01, 0x01, { SE_BUF(528), SE_DUMMY, SE_DUMMY } },
-    { 0xF0, 0x01, 0x00, 0x01, { SE_BUF(537), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(550), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(580), SE_DUMMY, SE_DUMMY } },
-    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(1088), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(599), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(608), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(649), SE_BUF(659), SE_DUMMY } },
-    { 0x90, 0x03, 0x00, 0x00, { SE_BUF(617), SE_BUF(629), SE_BUF(639) } },
-    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(669), SE_BUF(683), SE_DUMMY } },
-    { 0x90, 0x01, 0x01, 0x00, { SE_BUF(906), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x02, 0x00, 0x00, { SE_BUF(1352), SE_BUF(1366), SE_DUMMY } },
-    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(813), SE_DUMMY, SE_DUMMY } },
-    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(16), SE_DUMMY, SE_DUMMY } },
-    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(24), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(828), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x01, 0x00, { SE_BUF(856), SE_DUMMY, SE_DUMMY } },
-    { 0xA0, 0x01, 0x00, 0x00, { SE_BUF(338), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(54), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(72), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(845), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(87), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(105), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(440), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(116), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(125), SE_DUMMY, SE_DUMMY } },
-    { 0x70, 0x01, 0x01, 0x00, { SE_BUF(1556), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(1239), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(1226), SE_DUMMY, SE_DUMMY } },
-    { 0x10, 0x01, 0x00, 0x00, { SE_BUF(2305), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(878), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1208), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1217), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(1845), SE_BUF(1884), SE_DUMMY } },
-    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(915), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(165), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(192), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(1183), SE_DUMMY, SE_DUMMY } },
-    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(1199), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(275), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x02, 0x01, 0x00, { SE_BUF(943), SE_BUF(955), SE_DUMMY } },
-    { 0x70, 0x01, 0x01, 0x00, { SE_BUF(1831), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(2315), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(2339), SE_DUMMY, SE_DUMMY } },
-    { 0x01, 0x00, 0x00, 0x00, { SE_DUMMY, SE_DUMMY, SE_DUMMY } },
-    { 0x01, 0x00, 0x00, 0x00, { SE_DUMMY, SE_DUMMY, SE_DUMMY } },
-    { 0x01, 0x00, 0x00, 0x00, { SE_DUMMY, SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(239), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(2170), SE_DUMMY, SE_DUMMY } },
-    { 0x80, 0x02, 0x00, 0x00, { SE_BUF(2182), SE_BUF(2194), SE_DUMMY } },
-    { 0xFF, 0x01, 0x00, 0x00, { SE_BUF(2264), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(507), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(252), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(286), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x00, 0x00, { SE_BUF(295), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(306), SE_DUMMY, SE_DUMMY } },
-    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(315), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(324), SE_DUMMY, SE_DUMMY } },
-    { 0xA0, 0x03, 0x00, 0x00, { SE_BUF(353), SE_BUF(370), SE_BUF(388) } },
-    { 0x80, 0x02, 0x00, 0x00, { SE_BUF(1258), SE_BUF(1268), SE_DUMMY } },
-    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(771), SE_DUMMY, SE_DUMMY } },
-    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(780), SE_DUMMY, SE_DUMMY } },
-    { 0x80, 0x02, 0x00, 0x00, { SE_BUF(1278), SE_BUF(1293), SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(702), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(715), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(728), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(738), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(752), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(762), SE_DUMMY, SE_DUMMY } },
-    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(1935), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x01, 0x00, { SE_BUF(934), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(972), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(982), SE_DUMMY, SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(996), SE_DUMMY, SE_DUMMY } },
-    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(1248), SE_DUMMY, SE_DUMMY } },
-    { 0xFF, 0x03, 0x00, 0x00, { SE_BUF(1023), SE_BUF(1039), SE_BUF(1069) } },
-    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(2353), SE_DUMMY, SE_DUMMY } },
-    { 0xF0, 0x02, 0x00, 0x00, { SE_BUF(1571), SE_BUF(1585), SE_DUMMY } },
-    { 0x20, 0x01, 0x00, 0x00, { SE_BUF(1309), SE_DUMMY, SE_DUMMY } },
-    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(1328), SE_DUMMY, SE_DUMMY } },
-    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(1337), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1601), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x01, 0x00, 0x01, { SE_BUF(1626), SE_DUMMY, SE_DUMMY } },
-    { 0x80, 0x02, 0x00, 0x00, { SE_BUF(1640), SE_BUF(1651), SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1673), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x02, 0x01, 0x00, { SE_BUF(1694), SE_BUF(1706), SE_DUMMY } },
-    { 0x30, 0x02, 0x01, 0x00, { SE_BUF(1716), SE_BUF(1728), SE_DUMMY } },
-    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1747), SE_DUMMY, SE_DUMMY } },
-    { 0x30, 0x02, 0x01, 0x00, { SE_BUF(1756), SE_BUF(1768), SE_DUMMY } },
-    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1822), SE_DUMMY, SE_DUMMY } },
-    { 0xB0, 0x01, 0x00, 0x00, { SE_BUF(1808), SE_DUMMY, SE_DUMMY } },
-    { 0x40, 0x01, 0x00, 0x00, { SE_BUF(2086), SE_DUMMY, SE_DUMMY } },
-    { 0xB0, 0x01, 0x00, 0x00, { SE_BUF(2073), SE_DUMMY, SE_DUMMY } },
-    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(2053), SE_DUMMY, SE_DUMMY } },
-    { 0xB0, 0x01, 0x00, 0x00, { SE_BUF(2043), SE_DUMMY, SE_DUMMY } },
-    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(2097), SE_BUF(2111), SE_DUMMY } },
-    { 0xF0, 0x02, 0x00, 0x00, { SE_BUF(1946), SE_BUF(1963), SE_DUMMY } },
-    { 0xF0, 0x03, 0x00, 0x00, { SE_BUF(1981), SE_BUF(1991), SE_BUF(2001) } },
-    { 0xF0, 0x03, 0x00, 0x00, { SE_BUF(2278), SE_BUF(2287), SE_BUF(2296) } },
-    { 0x40, 0x02, 0x01, 0x00, { SE_BUF(2205), SE_BUF(2215), SE_DUMMY } },
-    { 0xF0, 0x02, 0x00, 0x00, { SE_BUF(2223), SE_BUF(2248), SE_DUMMY } },
-    { 0xB0, 0x03, 0x00, 0x00, { SE_BUF(2142), SE_BUF(2152), SE_BUF(2161) } },
-    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1737), SE_DUMMY, SE_DUMMY } }
-};
-
-/***** sd_wk.c (.data) *****/
-
-int spu_ch_tbl_800A2AC8[] = {
-    0x00000001,
-    0x00000001, 0x00000002, 0x00000004, 0x00000008,
-    0x00000010, 0x00000020, 0x00000040, 0x00000080,
-    0x00000100, 0x00000200, 0x00000400, 0x00000800,
-    0x00001000, 0x00002000, 0x00004000, 0x00008000,
-    0x00010000, 0x00020000, 0x00040000, 0x00080000,
-    0x00100000, 0x00200000, 0x00400000, 0x00800000
-};
-
-unsigned char blank_data_800A2B28[512] = {
-    0x0C,0x06, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x03, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0
-};
-
-unsigned char dummy_data_800A2D28[4096] = {
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0,
-    0x0C,0x02, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0, 0,0
+    /* pri,tracks,kind,character */
+    { 0x01, 0x01, 0x01, 0x00, { SE_DUMMY, SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(561), SE_DUMMY, SE_DUMMY }},
+    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(0), SE_DUMMY, SE_DUMMY }},
+    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(8), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(804), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(2122), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(2132), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1111), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1142), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(1159), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x00, 0x00, { SE_BUF(1171), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1376), SE_DUMMY, SE_DUMMY }},
+    { 0xB0, 0x01, 0x00, 0x00, { SE_BUF(1413), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1398), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1612), SE_DUMMY, SE_DUMMY }},
+    { 0xF0, 0x03, 0x00, 0x00, { SE_BUF(1428), SE_BUF(1441), SE_BUF(1452) }},
+    { 0x20, 0x01, 0x00, 0x00, { SE_BUF(406), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(430), SE_DUMMY, SE_DUMMY }},
+    { 0xA0, 0x01, 0x01, 0x01, { SE_BUF(1921), SE_DUMMY, SE_DUMMY }},
+    { 0xA0, 0x01, 0x01, 0x01, { SE_BUF(571), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x01, 0x00, { SE_BUF(693), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x00, 0x00, { SE_BUF(450), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(464), SE_BUF(476), SE_DUMMY }},
+    { 0x10, 0x01, 0x00, 0x00, { SE_BUF(498), SE_DUMMY, SE_DUMMY }},
+    { 0xA0, 0x01, 0x01, 0x01, { SE_BUF(518), SE_DUMMY, SE_DUMMY }},
+    { 0xA0, 0x01, 0x01, 0x01, { SE_BUF(528), SE_DUMMY, SE_DUMMY }},
+    { 0xF0, 0x01, 0x00, 0x01, { SE_BUF(537), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(550), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(580), SE_DUMMY, SE_DUMMY }},
+    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(1088), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(599), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(608), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(649), SE_BUF(659), SE_DUMMY }},
+    { 0x90, 0x03, 0x00, 0x00, { SE_BUF(617), SE_BUF(629), SE_BUF(639) }},
+    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(669), SE_BUF(683), SE_DUMMY }},
+    { 0x90, 0x01, 0x01, 0x00, { SE_BUF(906), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x02, 0x00, 0x00, { SE_BUF(1352), SE_BUF(1366), SE_DUMMY }},
+    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(813), SE_DUMMY, SE_DUMMY }},
+    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(16), SE_DUMMY, SE_DUMMY }},
+    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(24), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(828), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x01, 0x00, { SE_BUF(856), SE_DUMMY, SE_DUMMY }},
+    { 0xA0, 0x01, 0x00, 0x00, { SE_BUF(338), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(54), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(72), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(845), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(87), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(105), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(440), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(116), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(125), SE_DUMMY, SE_DUMMY }},
+    { 0x70, 0x01, 0x01, 0x00, { SE_BUF(1556), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(1239), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(1226), SE_DUMMY, SE_DUMMY }},
+    { 0x10, 0x01, 0x00, 0x00, { SE_BUF(2305), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(878), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1208), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1217), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(1845), SE_BUF(1884), SE_DUMMY }},
+    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(915), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(165), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(192), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(1183), SE_DUMMY, SE_DUMMY }},
+    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(1199), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(275), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x02, 0x01, 0x00, { SE_BUF(943), SE_BUF(955), SE_DUMMY }},
+    { 0x70, 0x01, 0x01, 0x00, { SE_BUF(1831), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(2315), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(2339), SE_DUMMY, SE_DUMMY }},
+    { 0x01, 0x00, 0x00, 0x00, { SE_DUMMY, SE_DUMMY, SE_DUMMY }},
+    { 0x01, 0x00, 0x00, 0x00, { SE_DUMMY, SE_DUMMY, SE_DUMMY }},
+    { 0x01, 0x00, 0x00, 0x00, { SE_DUMMY, SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(239), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(2170), SE_DUMMY, SE_DUMMY }},
+    { 0x80, 0x02, 0x00, 0x00, { SE_BUF(2182), SE_BUF(2194), SE_DUMMY }},
+    { 0xFF, 0x01, 0x00, 0x00, { SE_BUF(2264), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(507), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(252), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(286), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x00, 0x00, { SE_BUF(295), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(306), SE_DUMMY, SE_DUMMY }},
+    { 0x60, 0x01, 0x01, 0x00, { SE_BUF(315), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(324), SE_DUMMY, SE_DUMMY }},
+    { 0xA0, 0x03, 0x00, 0x00, { SE_BUF(353), SE_BUF(370), SE_BUF(388) }},
+    { 0x80, 0x02, 0x00, 0x00, { SE_BUF(1258), SE_BUF(1268), SE_DUMMY }},
+    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(771), SE_DUMMY, SE_DUMMY }},
+    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(780), SE_DUMMY, SE_DUMMY }},
+    { 0x80, 0x02, 0x00, 0x00, { SE_BUF(1278), SE_BUF(1293), SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(702), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(715), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(728), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(738), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(752), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x01, 0x01, 0x00, { SE_BUF(762), SE_DUMMY, SE_DUMMY }},
+    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(1935), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x01, 0x00, { SE_BUF(934), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(972), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(982), SE_DUMMY, SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(996), SE_DUMMY, SE_DUMMY }},
+    { 0x10, 0x01, 0x01, 0x00, { SE_BUF(1248), SE_DUMMY, SE_DUMMY }},
+    { 0xFF, 0x03, 0x00, 0x00, { SE_BUF(1023), SE_BUF(1039), SE_BUF(1069) }},
+    { 0x40, 0x01, 0x01, 0x00, { SE_BUF(2353), SE_DUMMY, SE_DUMMY }},
+    { 0xF0, 0x02, 0x00, 0x00, { SE_BUF(1571), SE_BUF(1585), SE_DUMMY }},
+    { 0x20, 0x01, 0x00, 0x00, { SE_BUF(1309), SE_DUMMY, SE_DUMMY }},
+    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(1328), SE_DUMMY, SE_DUMMY }},
+    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(1337), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1601), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x01, 0x00, 0x01, { SE_BUF(1626), SE_DUMMY, SE_DUMMY }},
+    { 0x80, 0x02, 0x00, 0x00, { SE_BUF(1640), SE_BUF(1651), SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1673), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x02, 0x01, 0x00, { SE_BUF(1694), SE_BUF(1706), SE_DUMMY }},
+    { 0x30, 0x02, 0x01, 0x00, { SE_BUF(1716), SE_BUF(1728), SE_DUMMY }},
+    { 0x20, 0x01, 0x01, 0x00, { SE_BUF(1747), SE_DUMMY, SE_DUMMY }},
+    { 0x30, 0x02, 0x01, 0x00, { SE_BUF(1756), SE_BUF(1768), SE_DUMMY }},
+    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1822), SE_DUMMY, SE_DUMMY }},
+    { 0xB0, 0x01, 0x00, 0x00, { SE_BUF(1808), SE_DUMMY, SE_DUMMY }},
+    { 0x40, 0x01, 0x00, 0x00, { SE_BUF(2086), SE_DUMMY, SE_DUMMY }},
+    { 0xB0, 0x01, 0x00, 0x00, { SE_BUF(2073), SE_DUMMY, SE_DUMMY }},
+    { 0x80, 0x01, 0x00, 0x00, { SE_BUF(2053), SE_DUMMY, SE_DUMMY }},
+    { 0xB0, 0x01, 0x00, 0x00, { SE_BUF(2043), SE_DUMMY, SE_DUMMY }},
+    { 0x90, 0x02, 0x00, 0x00, { SE_BUF(2097), SE_BUF(2111), SE_DUMMY }},
+    { 0xF0, 0x02, 0x00, 0x00, { SE_BUF(1946), SE_BUF(1963), SE_DUMMY }},
+    { 0xF0, 0x03, 0x00, 0x00, { SE_BUF(1981), SE_BUF(1991), SE_BUF(2001) }},
+    { 0xF0, 0x03, 0x00, 0x00, { SE_BUF(2278), SE_BUF(2287), SE_BUF(2296) }},
+    { 0x40, 0x02, 0x01, 0x00, { SE_BUF(2205), SE_BUF(2215), SE_DUMMY }},
+    { 0xF0, 0x02, 0x00, 0x00, { SE_BUF(2223), SE_BUF(2248), SE_DUMMY }},
+    { 0xB0, 0x03, 0x00, 0x00, { SE_BUF(2142), SE_BUF(2152), SE_BUF(2161) }},
+    { 0x90, 0x01, 0x00, 0x00, { SE_BUF(1737), SE_DUMMY, SE_DUMMY }}
 };
